requirements.txt
discord.py
websockets
aichttp
chardet